requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
click==8.1
glfw==2.7
imageio-ffmpeg==0.5
imageio==2.34
munch==4.0
numpy<2
opencv-contrib-python
pandas==2.2
pillow==10.3
plyfile==1.0
pydantic==2.8
pyopengl==3.1
requests==2.31
scipy==1.13
tqdm==4.66
PyYAML
${ADDON_PATH}/gaussian-splatting/submodules/diff-gaussian-rasterization
${ADDON_PATH}/gaussian-splatting/submodules/simple-knn
imgui_bundle==1.3.0
imgui==2.0.0
scikit-image==0.24
imagecodecs[all]==2024.6.1
pynvml
GPUtil